LayoutTests/fast/dom/Range/range-clone-contents-expected.txt

test contents:
| <div>
|   <a>
|     href="#"
|     "link"
| <div>
| <div>
|   <br>
|   <br>
|   <img>
|     id="img-tag"
|     src=""
| <div>
|   <br>
|   <br>
|   "text"

PASS: Cloning right complex subtree yields:
| <div>
|   <a>
|     href="#"
|     "link"
| <div>
| <div>
|   <br>
|   <br>

PASS: Cloning left complex subtree yields:
| <div>
| <div>
|   <br>
|   <br>
|   "text"

DONE
